ToDo-Backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
psycopg2-binary==2.9.10
